// src/app/models/enums.defs.ts
export enum StorageKeys {
    CartItems = 'cartItems'
}
export enum Currency {
    USD = 'USD',
    MXN = 'MXN',
    EUR = 'EUR'
}

export enum MovieCategory {
    Action = 'Action',
    Drama = 'Drama',
    Comedy = 'Comedy',
    ScienceFiction = 'Science Fiction',
    Mystery = 'Mystery',
    Biography = 'Biography',
    History = 'History',
    Romance = 'Romance',
    Crime = 'Crime',
    Thriller = 'Thriller'
}

export enum SnackCategory {
    Popcorn = 'Popcorn',
    Soda = 'Soda',
    Candy = 'Candy',
    Chocolate = 'Chocolate',
    Nachos = 'Nachos',
    Pretzels = 'Pretzels',
    IceCream = 'Ice Cream',
    HotDog = 'Hot Dog',
    Chips = 'Chips',
    FruitCup = 'Fruit Cup'
}

export enum ComboCategory {
    Combo1 = 'Soldado',
    Combo2 = 'EnPareja2',
    Combo3 = 'Triciclo'
}

export enum SeatingCol {
    A = "a",
    B = "b",
    C = "c",
    D = "d",
    E = "e",
    F = "f",
    G = "g",
    H = "h",
    I = "i",
    J = "j",
    K = "k",
    L = "l",
}